Migrate pre install checks to fluvio-cluster crate
