+++ src/cli/src/cluster/check.rs
@@ -18,6 +18,7 @@ use url::{Url};
 use structopt::StructOpt;
 use colored::*;
 use serde_json::{Value};
+use fluvio_cluster::ClusterInstaller;
 
 use crate::CliError;
 use super::*;
@@ -26,7 +27,6 @@ use super::*;
 const MIN_KUBE_VERSION: &str = "1.5.0";
 const DEFAULT_HELM_VERSION: &str = "3.2.0";
 const SYS_CHART_VERSION: &str = "0.2.0";
-const SYS_CHART_NAME: &str = "fluvio-sys";
 const DEFAULT_NAMESPACE: &str = "default";
 const DUMMY_LB_SERVICE: &str = "flv-dummy-service";
 const RESOURCE_SERVICE: &str = "service";
@@ -477,7 +477,15 @@ fn check_helm_version() -> Result<(), IoError> {
 }
 
 fn check_sys_charts() -> Result<(), IoError> {
-    let sys_charts = install::installed_sys_charts(SYS_CHART_NAME);
+    let sys_charts = ClusterInstaller::sys_charts().map_err(|err| {
+        IoError::new(
+            ErrorKind::Other,
+            format!(
+                "Error fetching installed fluvio system charts: {}",
+                err.to_string()
+            ),
+        )
+    })?;
     if sys_charts.len() == 1 {
         let installed_chart = sys_charts.first().unwrap();
         let installed_chart_version = installed_chart.app_version.clone();
