+++ src/cli/src/cluster/install/k8.rs
@@ -1,117 +1,115 @@
 use std::process::Command;
 use std::io::Error as IoError;
-use semver::Version;
-use k8_client::K8Config;
-use k8_config::KubeContext;
+// use semver::Version;
+// use k8_client::K8Config;
+// use k8_config::KubeContext;
 use std::io::ErrorKind;
-use std::net::IpAddr;
-use std::str::FromStr;
-use url::Url;
+// use std::net::IpAddr;
+// use std::str::FromStr;
+// use url::Url;
 
 use super::*;
 use fluvio_cluster::ClusterInstaller;
 
-fn get_cluster_server_host(kc_config: KubeContext) -> Result<String, IoError> {
-    if let Some(ctx) = kc_config.config.current_cluster() {
-        let server_url = ctx.cluster.server.to_owned();
-        let url = match Url::parse(&server_url) {
-            Ok(url) => url,
-            Err(e) => {
-                return Err(IoError::new(
-                    ErrorKind::Other,
-                    format!("error parsing server url {}", e.to_string()),
-                ))
-            }
-        };
-        Ok(url.host().unwrap().to_string())
-    } else {
-        Err(IoError::new(ErrorKind::Other, "no context found"))
-    }
-}
+// fn get_cluster_server_host(kc_config: KubeContext) -> Result<String, IoError> {
+//     if let Some(ctx) = kc_config.config.current_cluster() {
+//         let server_url = ctx.cluster.server.to_owned();
+//         let url = match Url::parse(&server_url) {
+//             Ok(url) => url,
+//             Err(e) => {
+//                 return Err(IoError::new(
+//                     ErrorKind::Other,
+//                     format!("error parsing server url {}", e.to_string()),
+//                 ))
+//             }
+//         };
+//         Ok(url.host().unwrap().to_string())
+//     } else {
+//         Err(IoError::new(ErrorKind::Other, "no context found"))
+//     }
+// }
 
-fn pre_install_check() -> Result<(), CliError> {
-    let helm_version = Command::new("helm")
-        .arg("version")
-        .arg("--short")
-        .output()
-        .map_err(|err| {
-            CliError::Other(format!(
-                "Helm package manager not found: {}",
-                err.to_string()
-            ))
-        })?;
-    let version_text = String::from_utf8(helm_version.stdout).unwrap();
-    let version_text_trimmed = &version_text[1..].trim();
+// fn pre_install_check() -> Result<(), CliError> {
+//     let helm_version = Command::new("helm")
+//         .arg("version")
+//         .arg("--short")
+//         .output()
+//         .map_err(|err| {
+//             CliError::Other(format!(
+//                 "Helm package manager not found: {}",
+//                 err.to_string()
+//             ))
+//         })?;
+//     let version_text = String::from_utf8(helm_version.stdout).unwrap();
+//     let version_text_trimmed = &version_text[1..].trim();
 
-    const DEFAULT_HELM_VERSION: &str = "3.2.0";
+//     const DEFAULT_HELM_VERSION: &str = "3.2.0";
 
-    if Version::parse(&version_text_trimmed) < Version::parse(DEFAULT_HELM_VERSION) {
-        return Err(CliError::Other(format!(
-            "Helm version {} is not compatible with fluvio platform, please install version >= {}",
-            version_text_trimmed, DEFAULT_HELM_VERSION
-        )));
-    }
+//     if Version::parse(&version_text_trimmed) < Version::parse(DEFAULT_HELM_VERSION) {
+//         return Err(CliError::Other(format!(
+//             "Helm version {} is not compatible with fluvio platform, please install version >= {}",
+//             version_text_trimmed, DEFAULT_HELM_VERSION
+//         )));
+//     }
 
-    const SYS_CHART_VERSION: &str = "";
-    const SYS_CHART_NAME: &str = "fluvio-sys";
+//     const SYS_CHART_VERSION: &str = "";
+//     const SYS_CHART_NAME: &str = "fluvio-sys";
 
-    let sys_charts = helm::installed_sys_charts(SYS_CHART_NAME);
-    if sys_charts.len() == 1 {
-        let installed_chart = sys_charts.first().unwrap();
-        let installed_chart_version = installed_chart.app_version.clone();
-        // checking version of chart found
-        if Version::parse(&installed_chart_version) < Version::parse(SYS_CHART_VERSION) {
-            return Err(CliError::Other(format!(
-                "Fluvio system chart {} is not compatible with fluvio platform, please install version >= {}",
-                installed_chart_version, SYS_CHART_VERSION
-            )));
-        }
-    } else if sys_charts.is_empty() {
-        return Err(CliError::Other(
-            "Fluvio system chart is not installed, please install fluvio-sys first".to_string(),
-        ));
-    } else {
-        return Err(CliError::Other(
-            "Multiple fluvio system charts found".to_string(),
-        ));
-    }
+//     let sys_charts = helm::installed_sys_charts(SYS_CHART_NAME);
+//     if sys_charts.len() == 1 {
+//         let installed_chart = sys_charts.first().unwrap();
+//         let installed_chart_version = installed_chart.app_version.clone();
+//         // checking version of chart found
+//         if Version::parse(&installed_chart_version) < Version::parse(SYS_CHART_VERSION) {
+//             return Err(CliError::Other(format!(
+//                 "Fluvio system chart {} is not compatible with fluvio platform, please install version >= {}",
+//                 installed_chart_version, SYS_CHART_VERSION
+//             )));
+//         }
+//     } else if sys_charts.is_empty() {
+//         return Err(CliError::Other(
+//             "Fluvio system chart is not installed, please install fluvio-sys first".to_string(),
+//         ));
+//     } else {
+//         return Err(CliError::Other(
+//             "Multiple fluvio system charts found".to_string(),
+//         ));
+//     }
 
-    let k8_config = K8Config::load()?;
+//     let k8_config = K8Config::load()?;
 
-    match k8_config {
-        K8Config::Pod(_) => {
-            // ignore server check for pod
-        }
-        K8Config::KubeConfig(config) => {
-            let server_host = match get_cluster_server_host(config) {
-                Ok(server) => server,
-                Err(e) => {
-                    return Err(CliError::Other(format!(
-                        "error fetching server from kube context {}",
-                        e.to_string()
-                    )))
-                }
-            };
-            if !server_host.trim().is_empty() {
-                if IpAddr::from_str(&server_host).is_ok() {
-                    return Err(CliError::Other(
-                        format!("Cluster in kube context cannot use IP address, please use minikube context: {}", server_host),
-                    ));
-                };
-            } else {
-                return Err(CliError::Other(
-                    "Cluster in kubectl context cannot have empty hostname".to_owned(),
-                ));
-            }
-        }
-    };
+//     match k8_config {
+//         K8Config::Pod(_) => {
+//             // ignore server check for pod
+//         }
+//         K8Config::KubeConfig(config) => {
+//             let server_host = match get_cluster_server_host(config) {
+//                 Ok(server) => server,
+//                 Err(e) => {
+//                     return Err(CliError::Other(format!(
+//                         "error fetching server from kube context {}",
+//                         e.to_string()
+//                     )))
+//                 }
+//             };
+//             if !server_host.trim().is_empty() {
+//                 if IpAddr::from_str(&server_host).is_ok() {
+//                     return Err(CliError::Other(
+//                         format!("Cluster in kube context cannot use IP address, please use minikube context: {}", server_host),
+//                     ));
+//                 };
+//             } else {
+//                 return Err(CliError::Other(
+//                     "Cluster in kubectl context cannot have empty hostname".to_owned(),
+//                 ));
+//             }
+//         }
+//     };
 
-    Ok(())
-}
+//     Ok(())
+// }
 
 pub async fn install_core(opt: InstallCommand) -> Result<(), CliError> {
-    pre_install_check().map_err(|err| CliError::Other(err.to_string()))?;
-
     let mut builder = ClusterInstaller::new()
         .with_namespace(opt.k8_config.namespace)
         .with_group_name(opt.k8_config.group_name)
