+++ src/cluster/Cargo.toml
@@ -15,6 +15,8 @@ path = "src/lib.rs"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+semver = "0.10.0"
+url = "2.1.1"
 tracing = "0.1.19"
 tracing-futures = "0.2.4"
 serde = "1.0.115"
